Fix pymedusa/SickRage/issues/1078 (#1079)

+++ sickbeard/tv.py
@@ -1334,8 +1334,9 @@ class TVShow(TVObject):
                                            (id=self.indexerid, related_file=related_file, error_msg=e), logger.WARNING)
 
         # clean up any empty season folders after deletion of associated files
-        for sub_dir in ek(os.listdir, self.location):
-            helpers.delete_empty_folders(ek(os.path.join, self.location, sub_dir), self.location)
+        if ek(os.path.isdir, self.location):
+            for sub_dir in ek(os.listdir, self.location):
+                    helpers.delete_empty_folders(ek(os.path.join, self.location, sub_dir), self.location)
 
         if sql_l:
             main_db_con = db.DBConnection()
